refatoring the action bar components
